Refactored EventPythonModule and EventSystem
